Backend/requirements.txt
django==4.2.6
psycopg2-binary==2.9.9
Pillow==10.0.0
djangorestframework==3.14.0
django-phonenumber-field==7.2.0
phonenumbers==8.13.24
psycopg-binary==3.1.12
flake8==6.1.0
black==23.10.1
isort==5.12.0
django-filter==23.3
django-autoslug==1.9.9

